reorganization part1
