+++ client/src/components/application_page.js
@@ -3,7 +3,7 @@ import ExistingBookmarks from './existingbookmarks';
 import ExistingFolders from './existingfolders';
 import AllReminderList from './all_of_reminder_list';
 import './applicationcss.css';
-import ReminderListRedux from './reminder_list_redux';
+import ReminderListRedux from './reminders/reminder_list_redux';
 
 const ApplicationPage = () => (
