+++ client/src/components/reminders/reminder_list.css
@@ -7,22 +7,21 @@
     background-color: lightgray;
     list-style-type: none;
 }
+
 .listItemElements{
     font-size: 1rem;
     width: 100%;
     height: 13%;
     color: black;
-    background-image: url("../images/RirstColorBackground.jpg");
-
+    background-image: url("./images/light-reminder-bg.jpg");
 }
+
 .allListItemElements{
     list-style-type: none;
     height: 90vh;
     margin: 0;
     overflow: auto;
-
 }
-
 
 .iconStyle{
     width: 20%;
@@ -31,28 +30,27 @@
     top: 10%;
     left: 3%;
 }
+
 .partOfList{
     display: inline-block;
     position: relative;
 
 }
+
 .listStyle{
     position: relative;
     left: 7%;
     vertical-align: top;
     width: 60%;
     top: 20%;
-
-
-
 }
+
 .btnList{
     position: relative;
     font-size: .7rem;
     width: 10%;
     left: 9%;
     top: 9%;
-
 }
 
 .inLine{
@@ -66,6 +64,7 @@
     position: relative;
     left: 20%;
 }
+
 .titleClass{
     position: relative;
     width: 65%;
@@ -74,9 +73,8 @@
     white-space: pre;
     height: 20%;
     color: black;
-
-
 }
+
 .insideDateClass{
     font-size: .8rem;
     margin: 0;
