+++ client/src/components/team_page/teampage.css
@@ -1,6 +1,6 @@
 
 .teampage-background {
-    background-image: url(./images/team-page-bg.jpg);
+    background-image: url(../team_page//images/team-page-bg.jpg);
     background-size: contain;
     background-position: center;
     background-repeat: repeat;
